feat: allow the initial offset to be set in the starting phase
This commit allows the initial offset for SinglePartitionConsumer
to be set in the starting phase. This allows the service to backoff
and retry when fetching the initial offset outside of the New
function.

+++ pkg/kafkav2/offset.go
@@ -16,8 +16,8 @@ import (
 const (
 	// Special offsets in Kafka that refer to the start or end offset for
 	// a partition.
-	OffsetStart = -2
-	OffsetEnd   = -1
+	OffsetStart = int64(-2)
+	OffsetEnd   = int64(-1)
 )
 
 type Committer struct {
